rename + prevent map exception
